Re-enable table and filters for cancer biomarkers
