+++ src/components/CancerBiomarkers/Detail.js
@@ -1,12 +1,9 @@
-import React, { Component, Fragment } from 'react';
+import React from 'react';
 import { Query } from 'react-apollo';
 import gql from 'graphql-tag';
-import crossfilter from 'crossfilter2';
-import * as d3 from 'd3';
-import dc from 'dc';
-import { OtTable } from 'ot-ui';
+import { Typography } from '@material-ui/core';
 
-import DCContainer from '../DCContainer';
+import FilterTable from './FilterTable';
 
 const query = gql`
   query CancerBiomarkersQuery($ensgId: String!) {
@@ -33,235 +30,20 @@ const query = gql`
   }
 `;
 
-const columns = [
-  { id: 'biomarker', label: 'Biomarker' },
-  {
-    id: 'diseases',
-    label: 'Disease',
-    renderCell: rowData => {
-      return rowData.diseases.map(disease => disease.name).join(', ');
-    },
-  },
-  { id: 'drugName', label: 'Drug' },
-  { id: 'associationType', label: 'Association' },
-  { id: 'evidenceLevel', label: 'Evidence' },
-  {
-    id: 'sources',
-    label: 'Sources',
-    renderCell: rowData => {
-      const result = rowData.sources.map(source => (
-        <a
-          key={source.url}
-          href={source.url}
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          {source.name}
-        </a>
-      ));
-      return result;
-    },
-  },
-];
+const CancerBiomarkersDetail = ({ ensgId, symbol }) => {
+  return (
+    <React.Fragment>
+      <Typography variant="h5">{symbol} - Cancer Biomarkers</Typography>
+      <Query query={query} variables={{ ensgId }}>
+        {({ loading, error, data }) => {
+          if (loading || error) return null;
 
-class CancerBiomarkersDetails extends Component {
-  state = {
-    filteredRows: [],
-  };
+          const { rows } = data.target.details.cancerBiomarkers;
+          return <FilterTable rows={rows} />;
+        }}
+      </Query>
+    </React.Fragment>
+  );
+};
 
-  componentDidMount() {
-    this.renderCharts();
-  }
-
-  render() {
-    const { filteredRows } = this.state;
-
-    return (
-      <Fragment>
-        <DCContainer id="biomarkers-by-drug" title="Biomarkers by drug" />
-        <DCContainer
-          id="biomarkers-by-association"
-          title="Biomarkers by association"
-        />
-        <DCContainer
-          id="biomarkers-by-evidence"
-          title="Biomarkers by Evidence"
-        />
-        <DCContainer
-          id="biomarkers-drugs-heatmap"
-          title="Biomarkers and drugs"
-        />
-        <OtTable columns={columns} data={filteredRows} />;
-      </Fragment>
-    );
-  }
-
-  renderCharts() {
-    const { rows } = this.props;
-    // TODO: Remove the following to re-enable detail
-    if (!rows) return;
-
-    const biomarkers = crossfilter(rows);
-
-    const biomarkersByDrug = biomarkers.dimension(d => d.drugName);
-    const biomarkersByAssociation = biomarkers.dimension(
-      d => d.associationType
-    );
-    const biomarkersByEvidence = biomarkers.dimension(d => d.evidenceLevel);
-    const biomarkerAndDrug = biomarkers.dimension(d => [
-      d.biomarker,
-      d.drugName,
-    ]);
-
-    const biomarkersByDrugGroup = biomarkersByDrug.group().reduce(
-      (acc, data) => {
-        if (data.biomarker in acc) {
-          acc[data.biomarker]++;
-        } else {
-          acc[data.biomarker] = 1;
-        }
-        return acc;
-      },
-      (acc, data) => {
-        acc[data.biomarker]--;
-        if (acc[data.biomarker] === 0) {
-          delete acc[data.biomarker];
-        }
-        return acc;
-      },
-      () => ({})
-    );
-
-    const biomarkersByAssociationGroup = biomarkersByAssociation.group().reduce(
-      (acc, data) => {
-        if (data.biomarker in acc) {
-          acc[data.biomarker]++;
-        } else {
-          acc[data.biomarker] = 1;
-        }
-        return acc;
-      },
-      (acc, data) => {
-        acc[data.biomarker]--;
-        if (acc[data.biomarker] === 0) {
-          delete acc[data.biomarker];
-        }
-        return acc;
-      },
-      () => ({})
-    );
-
-    const biomarkersByEvidenceGroup = biomarkersByEvidence.group().reduce(
-      (acc, data) => {
-        if (data.biomarker in acc) {
-          acc[data.biomarker]++;
-        } else {
-          acc[data.biomarker] = 1;
-        }
-        return acc;
-      },
-      (acc, data) => {
-        acc[data.biomarker]--;
-        if (acc[data.biomarker] === 0) {
-          delete acc[data.biomarker];
-        }
-        return acc;
-      },
-      () => ({})
-    );
-
-    const biomarkerAndDrugGroup = biomarkerAndDrug.group().reduceCount();
-
-    const biomarkersByDrugChart = dc.rowChart('#biomarkers-by-drug');
-    const biomarkersByAssociationChart = dc.pieChart(
-      '#biomarkers-by-association'
-    );
-    const biomarkersByEvidenceChart = dc.pieChart('#biomarkers-by-evidence');
-    const biomarkersDrugsHeatmap = dc.heatMap('#biomarkers-drugs-heatmap');
-
-    biomarkersByDrugChart
-      .width(280)
-      .height(280)
-      .margins({ top: 20, left: 10, right: 10, bottom: 20 })
-      .label(d => d.key)
-      .valueAccessor(d => Object.keys(d.value).length)
-      .group(biomarkersByDrugGroup)
-      .dimension(biomarkersByDrug)
-      .xAxis();
-
-    biomarkersByDrugChart.render();
-
-    biomarkersByDrugChart.on('filtered', d => {
-      this.setState({ filteredRows: biomarkers.allFiltered() });
-    });
-
-    biomarkersByAssociationChart
-      .width(280)
-      .height(280)
-      .radius(120)
-      .label(d => d.key)
-      .valueAccessor(d => Object.keys(d.value).length)
-      .group(biomarkersByAssociationGroup)
-      .dimension(biomarkersByAssociation)
-      .render();
-
-    biomarkersByAssociationChart.on('filtered', d => {
-      this.setState({ filteredRows: biomarkers.allFiltered() });
-    });
-
-    biomarkersByEvidenceChart
-      .width(280)
-      .height(280)
-      .radius(120)
-      .label(d => d.key)
-      .valueAccessor(d => Object.keys(d.value).length)
-      .group(biomarkersByEvidenceGroup)
-      .dimension(biomarkersByEvidence)
-      .render();
-
-    biomarkersByEvidenceChart.on('filtered', d => {
-      this.setState({ filteredRows: biomarkers.allFiltered() });
-    });
-
-    biomarkersDrugsHeatmap
-      .width(380)
-      .height(450)
-      .margins({ top: 20, left: 200, right: 10, bottom: 150 })
-      .dimension(biomarkerAndDrug)
-      .group(biomarkerAndDrugGroup)
-      .keyAccessor(d => d.key[0])
-      .valueAccessor(d => d.key[1])
-      .colorAccessor(d => {
-        return d.value > 0 ? 1 : 0;
-      })
-      .colors(
-        d3
-          .scaleOrdinal()
-          .domain([0, 1])
-          .range(['#ccc', '#3182bd'])
-      )
-      .xBorderRadius(0)
-      .yBorderRadius(0)
-      .render();
-
-    biomarkersDrugsHeatmap.on('filtered', d => {
-      this.setState({ filteredRows: biomarkers.allFiltered() });
-    });
-
-    this.setState({
-      filteredRows: biomarkers.allFiltered(),
-    });
-
-    biomarkersDrugsHeatmap
-      .selectAll('g.cols.axis > text')
-      .attr('transform', function(d) {
-        const coord = this.getBBox();
-        const x = coord.x + coord.width / 2;
-        const y = coord.y + coord.height / 2;
-        return `rotate(-45 ${x} ${y})`;
-      })
-      .style('text-anchor', 'end');
-  }
-}
-
-export default CancerBiomarkersDetails;
+export default CancerBiomarkersDetail;
